Update simple test to check if JNI+Rust works
